exla/models/robotpoint/_implementations/robotpoint_cpu.py
"""
RoboPoint model implementation for CPU.
"""

import os
import sys
import time
import random
import numpy as np
from pathlib import Path
from typing import Dict, List, Tuple, Union, Optional

try:
    from PIL import Image, ImageDraw
except ImportError:
    print("Error: PIL not found. Please install it with: pip install pillow")
    sys.exit(1)

from ..._base import BaseModel


class RoboPointCPU(BaseModel):
    """
    RoboPoint model implementation for CPU.
    This is a mock implementation for demonstration purposes.
    """

    def __init__(self):
        """Initialize the RoboPoint model for CPU."""
        super().__init__()
        self.name = "RoboPoint"
        self.device = "cpu"
        self._model = None
        self._is_loaded = False
        self._install_dependencies()
        self._load_model()

    def _install_dependencies(self) -> None:
        """Install required dependencies."""
        try:
            # In a real implementation, we would install dependencies here
            # For the mock, we'll just simulate a delay
            time.sleep(1)
            print("✓ [1.0s] Installing dependencies...")
            print("✓ [1.0s] Dependencies installed successfully")
        except Exception as e:
            print(f"Error installing dependencies: {str(e)}")
            raise

    def _load_model(self) -> None:
        """Load the RoboPoint model."""
        try:
            # In a real implementation, we would load the model here
            # For the mock, we'll just simulate a delay
            time.sleep(1)
            self._model = "mock_model"
            self._is_loaded = True
            print("✓ [1.0s] Loading RoboPoint model on CPU...")
        except Exception as e:
            print(f"Error loading model: {str(e)}")
            raise

    def load_image(self, image_path: str) -> Image.Image:
        """
        Load an image from the given path.
        
        Args:
            image_path: Path to the image file.
            
        Returns:
            PIL Image object.
        """
        try:
            if not os.path.exists(image_path):
                raise FileNotFoundError(f"Image not found: {image_path}")
            
            image = Image.open(image_path).convert("RGB")
            return image
        except Exception as e:
            print(f"Error loading image: {str(e)}")
            raise

    def predict_keypoints(self, image: Image.Image, instruction: str) -> List[Dict[str, Union[float, str]]]:
        """
        Predict keypoints on the image based on the instruction.
        
        Args:
            image: PIL Image object.
            instruction: Text instruction for the model.
            
        Returns:
            List of dictionaries containing keypoint information.
        """
        if not self._is_loaded:
            raise RuntimeError("Model is not loaded. Call _load_model() first.")
        
        # In a real implementation, we would run inference here
        # For the mock, we'll generate random keypoints
        width, height = image.size
        num_points = random.randint(3, 8)
        
        keypoints = []
        for i in range(num_points):
            x = random.uniform(0.1, 0.9) * width
            y = random.uniform(0.1, 0.9) * height
            confidence = random.uniform(0.7, 0.99)
            
            keypoints.append({
                "x": x,
                "y": y,
                "confidence": confidence,
                "label": f"Point {i+1}"
            })
        
        return keypoints

    def visualize(self, image: Image.Image, keypoints: List[Dict[str, Union[float, str]]], output_path: str) -> None:
        """
        Visualize the keypoints on the image and save to the output path.
        
        Args:
            image: PIL Image object.
            keypoints: List of dictionaries containing keypoint information.
            output_path: Path to save the visualization.
        """
        try:
            # Create a copy of the image for visualization
            vis_image = image.copy()
            draw = ImageDraw.Draw(vis_image)
            
            # Draw keypoints
            for i, kp in enumerate(keypoints):
                x, y = kp["x"], kp["y"]
                label = kp["label"]
                confidence = kp["confidence"]
                
                # Draw a circle for each keypoint
                radius = 10
                draw.ellipse([(x-radius, y-radius), (x+radius, y+radius)], 
                             fill=(255, 0, 0), outline=(0, 0, 0))
                
                # Draw the label
                draw.text((x+radius+5, y-radius), 
                          f"{label} ({confidence:.2f})", 
                          fill=(0, 0, 0))
            
            # Save the visualization
            os.makedirs(os.path.dirname(output_path), exist_ok=True)
            vis_image.save(output_path)
            print(f"Visualization saved to: {output_path}")
        except Exception as e:
            print(f"Error visualizing keypoints: {str(e)}")
            raise 