exla/models/robotpoint/_implementations/requirements/requirements_cpu.txt
torch>=2.0.0
transformers>=4.31.0
pillow>=9.0.0
numpy>=1.24.0
huggingface_hub>=0.16.4 